requirements.txt
numpy
scipy
pydantic
requests
python-dotenv
webrtcvad
pydub
soundfile
librosa